+++ docs/.eslintrc.cjs
@@ -37,6 +37,7 @@ module.exports = {
     {
       files: ['next-env.d.ts'],
       rules: {
+        'prettier/prettier': 'off',
         '@typescript-eslint/triple-slash-reference': 'off',
       },
     },