+++ packages/source-duckdb/e2e/utils/create-duckdb-e2e-memory-db.ts
@@ -6,7 +6,7 @@ export const createDuckDBE2EMemoryDb = async (): Promise<DuckDBConnection> => {
   const instance = await DuckDBInstance.create(':memory:', {
     access_mode: 'READ_WRITE',
     max_memory: '64MB',
-    threads: `${Math.min(os.availableParallelism() - 1, 4)}`,
+    threads: `${Math.max(1, Math.min(os.availableParallelism() - 1, 4))}`,
   });
   return await instance.connect();
 };