+++ packages/source-duckdb/vitest.config.ts
@@ -13,7 +13,7 @@ const cspeed = isCodeSpeedEnabled ? codspeedPlugin() : undefined;
 
 export default defineConfig({
   esbuild: {
-    target: ['node18'],
+    target: ['node20'],
   },
   plugins: [tsconfigPaths(), ...[cspeed].filter(Boolean)],
   cacheDir: '../../.cache/vite/source-duckdb',
@@ -22,7 +22,6 @@ export default defineConfig({
     // @link https://vitest.dev/config/#clearmocks
     clearMocks: true,
     coverage: {
-      all: true,
       include: ['src/**/*.{js,jsx,ts,tsx}'],
       provider: 'istanbul',
       reporter: ['text', 'json', 'clover'],